+++ src/core/BUILD.gn
@@ -393,6 +393,7 @@ openthread_core_files = [
   "common/binary_search.cpp",
   "common/binary_search.hpp",
   "common/bit_set.hpp",
+  "common/bit_utils.hpp",
   "common/callback.hpp",
   "common/clearable.hpp",
   "common/const_cast.hpp",
