+++ src/core/common/num_utils.hpp
@@ -34,7 +34,6 @@
 #ifndef NUM_UTILS_HPP_
 #define NUM_UTILS_HPP_
 
-#include "common/encoding.hpp"
 #include "common/numeric_limits.hpp"
 #include "common/type_traits.hpp"
 
@@ -254,259 +253,6 @@ template <typename IntType> inline IntType DivideAndRoundUp(IntType aDividend, I
  * @returns The @p aUint32 value as `unsigned long`.
  */
 inline unsigned long ToUlong(uint32_t aUint32) { return static_cast<unsigned long>(aUint32); }
-
-/**
- * Counts the number of `1` bits in the binary representation of a given unsigned int bit-mask value.
- *
- * @tparam UintType   The unsigned int type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- *
- * @param[in] aMask   A bit mask.
- *
- * @returns The number of `1` bits in @p aMask.
- */
-template <typename UintType> uint8_t CountBitsInMask(UintType aMask)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    uint8_t count = 0;
-
-    while (aMask != 0)
-    {
-        aMask &= aMask - 1;
-        count++;
-    }
-
-    return count;
-}
-
-/**
- * Sets the specified bit of the given integer to 1.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- *
- * @param[in,out]  aBits       The integer to set the bit.
- * @param[in]      aBitOffset  The bit offset to set. The bit offset starts with zero corresponding to the
- *                             least-significant bit.
- */
-template <typename UintType> void SetBit(UintType &aBits, uint8_t aBitOffset)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    aBits |= static_cast<UintType>(static_cast<UintType>(1) << aBitOffset);
-}
-
-/**
- * Clears the specified bit of the given integer.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- *
- * @param[in,out]  aBits       The integer to clear the bit.
- * @param[in]      aBitOffset  The bit offset to clear. The bit offset starts with zero corresponding to the
- *                             least-significant bit.
- */
-template <typename UintType> void ClearBit(UintType &aBits, uint8_t aBitOffset)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    aBits &= (~(static_cast<UintType>(1) << aBitOffset));
-}
-
-/**
- * Gets the value of the specified bit of the given integer.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- *
- * @param[in] aBits       The integer to get the bit.
- * @param[in] aBitOffset  The bit offset to get. The bit offset starts with zero corresponding to the
- *                        least-significant bit.
- *
- * @returns The value of the specified bit.
- */
-template <typename UintType> bool GetBit(UintType aBits, uint8_t aBitOffset)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    return (aBits & (static_cast<UintType>(1) << aBitOffset)) != 0;
-}
-
-/**
- * Writes the specified bit of the given integer to the given value (0 or 1).
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- *
- * @param[in,out]  aBits      The integer to write the bit.
- * @param[in]      aBitOffset The bit offset to get. The bit offset starts with zero corresponding to the
- *                            least-significant bit.
- * @param[in]      aValue     The value to write.
- */
-template <typename UintType> void WriteBit(UintType &aBits, uint8_t aBitOffset, bool aValue)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    if (aValue)
-    {
-        SetBit<UintType>(aBits, aBitOffset);
-    }
-    else
-    {
-        ClearBit<UintType>(aBits, aBitOffset);
-    }
-}
-
-/**
- * Gets the offset of the lowest non-zero bit in the given mask.
- *
- * @tparam UintType  The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- *
- * @param[in] aMask  The mask (MUST not be 0) to calculate the offset of the lowest non-zero bit.
- *
- * @returns The offset of the lowest non-zero bit in the mask.
- */
-template <typename UintType> inline constexpr uint8_t BitOffsetOfMask(UintType aMask)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    return (aMask & 0x1) ? 0 : (1 + BitOffsetOfMask<UintType>(aMask >> 1));
-}
-
-/**
- * Writes the specified bits of the given integer to the given value.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- * @tparam kMask      The bit mask (MUST not be 0) to write. The @p kMask must be provided in a shifted form.
- * @tparam kOffset    The bit offset to write. The default @p kOffset is computed from the given @p kMask.
- *
- * @param[in,out]  aBits   The integer to write the bits.
- * @param[in]      aValue  The value to write.
- */
-template <typename UintType, UintType kMask, UintType kOffset = BitOffsetOfMask(kMask)>
-void WriteBits(UintType &aBits, UintType aValue)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    aBits = ((aBits & ~kMask) | ((aValue << kOffset) & kMask));
-}
-
-/**
- * Writes the specified bits of the given integer to the given value and returns the updated integer.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- * @tparam kMask      The bit mask (MUST not be 0) to write. The @p kMask must be provided in a shifted form.
- * @tparam kOffset    The bit offset to write. The default @p kOffset is computed from the given @p kMask.
- *
- * @param[in] aBits   The integer to write the bits.
- * @param[in] aValue  The value to write.
- *
- * @returns The updated integer.
- */
-template <typename UintType, UintType kMask, UintType kOffset = BitOffsetOfMask(kMask)>
-UintType UpdateBits(UintType aBits, UintType aValue)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    return ((aBits & ~kMask) | ((aValue << kOffset) & kMask));
-}
-
-/**
- * Read the value of the specified bits of the given integer.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- * @tparam kMask      The bit mask (MUST not be 0) to write. The @p kMask must be provided in a shifted form.
- * @tparam kOffset    The bit offset to write. The default @p kOffset is computed from the given @p kMask.
- *
- * @param[in] aBits   The integer to read the bits.
- *
- * @returns The value of the specified bits.
- */
-template <typename UintType, UintType kMask, UintType kOffset = BitOffsetOfMask(kMask)>
-UintType ReadBits(UintType aBits)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    return (aBits & kMask) >> kOffset;
-}
-
-/**
- * Writes the specified bits of the given integer stored in little-endian format to the given value and returns the
- * updated integer stored in little-endian format.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- * @tparam kMask      The bit mask (MUST not be 0) to write. The @p kMask must be provided in a shifted form.
- * @tparam kOffset    The bit offset to write. The default @p kOffset is computed from the given @p kMask.
- *
- * @param[in]  aBits   The integer to write the bits.
- * @param[in]  aValue  The value to write.
- *
- * @returns The updated integer.
- */
-template <typename UintType, UintType kMask, UintType kOffset = BitOffsetOfMask(kMask)>
-UintType UpdateBitsLittleEndian(UintType aBits, UintType aValue)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    return LittleEndian::HostSwap<UintType>((LittleEndian::HostSwap<UintType>(aBits) & ~kMask) |
-                                            ((aValue << kOffset) & kMask));
-}
-
-/**
- * Writes the specified bits of the given integer stored in big-endian format to the given value and returns the updated
- * integer stored in big-endian format.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- * @tparam kMask      The bit mask (MUST not be 0) to write. The @p kMask must be provided in a shifted form.
- * @tparam kOffset    The bit offset to write. The default @p kOffset is computed from the given @p kMask.
- *
- * @param[in]  aBits   A pointer to the integer to write the bits.
- * @param[in]  aValue  The value to write.
- *
- * @returns The updated integer.
- */
-template <typename UintType, UintType kMask, UintType kOffset = BitOffsetOfMask(kMask)>
-UintType UpdateBitsBigEndian(UintType aBits, UintType aValue)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    return BigEndian::HostSwap<UintType>((BigEndian::HostSwap<UintType>(aBits) & ~kMask) |
-                                         ((aValue << kOffset) & kMask));
-}
-
-/**
- * Read the value of the specified bits of the given integer stored in little-endian format.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- * @tparam kMask      The bit mask (MUST not be 0) to write. The @p kMask must be provided in a shifted form.
- * @tparam kOffset    The bit offset to write. The default @p kOffset is computed from the given @p kMask.
- *
- * @param[in] aBits   The integer stored in little-endian format to read the bits.
- *
- * @returns The value of the specified bits.
- */
-template <typename UintType, UintType kMask, UintType kOffset = BitOffsetOfMask(kMask)>
-UintType ReadBitsLittleEndian(UintType aBits)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    return (LittleEndian::HostSwap<UintType>(aBits) & kMask) >> kOffset;
-}
-
-/**
- * Read the value of the specified bits of the given integer stored in big-endian format.
- *
- * @tparam UintType   The value type (MUST be `uint8_t`, `uint16_t`, `uint32_t`, or `uint64_t`).
- * @tparam kMask      The bit mask (MUST not be 0) to write. The @p kMask must be provided in a shifted form.
- * @tparam kOffset    The bit offset to write. The default @p kOffset is computed from the given @p kMask.
- *
- * @param[in] aBits   The integer stored in big-endian format to read the bits.
- *
- * @returns The value of the specified bits.
- */
-template <typename UintType, UintType kMask, UintType kOffset = BitOffsetOfMask(kMask)>
-UintType ReadBitsBigEndian(UintType aBits)
-{
-    static_assert(TypeTraits::IsUint<UintType>::kValue, "UintType must be an unsigned int (8, 16, 32, or 64 bit len)");
-
-    return (BigEndian::HostSwap<UintType>(aBits) & kMask) >> kOffset;
-}
 
 } // namespace ot
